feat: Upgrade agent to support linux-akri OTA's
- Linux-Akri requires a docker image (instead of a firmware image) to
update. Therefore, updated agent to detect which OTA process will be
followed, based on given `NEW_FIRMWARE_PATH` or `CONTANER_IMG`. In ca-
se of the Linux-Akri update, the image is transmitted to the client.
